+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/entities/cms/CmsLottery.java
@@ -34,7 +34,7 @@ public class CmsLottery implements java.io.Serializable {
     private Integer id;
     @GeneratorColumn(title = "站点", condition = true)
     @JsonIgnore
-    private int siteId;
+    private short siteId;
     @GeneratorColumn(title = "开始日期", condition = true, order = true)
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private Date startDate;
@@ -69,7 +69,7 @@ public class CmsLottery implements java.io.Serializable {
     public CmsLottery() {
     }
 
-    public CmsLottery(int siteId, Date startDate, Date endDate, int intervalHour, int gift, int totalGift, int lastGift,
+    public CmsLottery(short siteId, Date startDate, Date endDate, int intervalHour, int gift, int totalGift, int lastGift,
             int lotteryCount, int fractions, int numerator, String title, boolean disabled) {
         this.siteId = siteId;
         this.startDate = startDate;
@@ -85,7 +85,7 @@ public class CmsLottery implements java.io.Serializable {
         this.disabled = disabled;
     }
 
-    public CmsLottery(int siteId, Date startDate, Date endDate, int intervalHour, int gift, int totalGift, int lastGift,
+    public CmsLottery(short siteId, Date startDate, Date endDate, int intervalHour, int gift, int totalGift, int lastGift,
             int lotteryCount, int fractions, int numerator, String url, String title, String description, boolean disabled,
             Integer extendId) {
         this.siteId = siteId;
@@ -118,11 +118,11 @@ public class CmsLottery implements java.io.Serializable {
     }
 
     @Column(name = "site_id", nullable = false)
-    public int getSiteId() {
+    public short getSiteId() {
         return this.siteId;
     }
 
-    public void setSiteId(int siteId) {
+    public void setSiteId(short siteId) {
         this.siteId = siteId;
     }
