+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/entities/log/LogUpload.java
@@ -33,7 +33,7 @@ public class LogUpload implements java.io.Serializable {
     private Long id;
     @GeneratorColumn(title = "站点", condition = true)
     @JsonIgnore
-    private int siteId;
+    private short siteId;
     @GeneratorColumn(title = "用户", condition = true)
     private long userId;
     @GeneratorColumn(title = "渠道", condition = true)
@@ -52,7 +52,7 @@ public class LogUpload implements java.io.Serializable {
     public LogUpload() {
     }
 
-    public LogUpload(int siteId, long userId, String channel, boolean image, long fileSize, Date createDate, String filePath) {
+    public LogUpload(short siteId, long userId, String channel, boolean image, long fileSize, Date createDate, String filePath) {
         this.siteId = siteId;
         this.userId = userId;
         this.channel = channel;
@@ -62,7 +62,7 @@ public class LogUpload implements java.io.Serializable {
         this.filePath = filePath;
     }
 
-    public LogUpload(int siteId, long userId, String channel, boolean image, long fileSize, String ip, Date createDate,
+    public LogUpload(short siteId, long userId, String channel, boolean image, long fileSize, String ip, Date createDate,
             String filePath) {
         this.siteId = siteId;
         this.userId = userId;
@@ -87,11 +87,11 @@ public class LogUpload implements java.io.Serializable {
     }
 
     @Column(name = "site_id", nullable = false)
-    public int getSiteId() {
+    public short getSiteId() {
         return this.siteId;
     }
 
-    public void setSiteId(int siteId) {
+    public void setSiteId(short siteId) {
         this.siteId = siteId;
     }
